Add files via upload
This code reads the CSV file and groups the data by user_location, then counts the number of tweets from each location and creates a horizontal bar chart of the top 10 tweeting locations. You can customize this code to create different types of visualizations based on the data in your CSV file.